fix(stream): recreate consumer when a stream instance is restarted
This reverts commit f0dca3a45738cecc675aa0cae6440c3affe11e13.
